Move search function to downloader file.
