+++ src/download/exchangeDownloader.ts
@@ -10,6 +10,8 @@ import { writeFileSync, ensureDirSync } from "fs-extra";
 import fetch, { Response } from "node-fetch";
 import path from "path";
 
+import { getBearer } from "./bearerToken";
+import { removeVersionSpecificInformation } from "./exchangeTools";
 import {
   RawRestApi,
   RestApi,
@@ -202,4 +204,33 @@ export async function getSpecificApi(
   if (!version) return null;
   const api = await getAsset(accessToken, `${groupId}/${assetId}/${version}`);
   return api ? convertResponseToRestApi(api) : null;
+}
+
+/**
+ * Gets information about all the APIs from exchange that match the given search
+ * string for the version deployed in the given environment.
+ * If it fails to get information about the deployed version of an API, it
+ * removes all the version specific information from the returned object.
+ *
+ * @param search - Exchange search query
+ * @param deployment - RegExp matching the desired deployment targets
+ *
+ * @returns Information about the APIs found.
+ */
+export async function search(
+  search: string,
+  deployment: RegExp
+): Promise<RestApi[]> {
+  const token = await getBearer(
+    process.env.ANYPOINT_USERNAME,
+    process.env.ANYPOINT_PASSWORD
+  );
+  const apis = await searchExchange(token, search);
+  const promises = apis.map(async api => {
+    const version = await getVersionByDeployment(token, api, deployment);
+    return version
+      ? getSpecificApi(token, api.groupId, api.assetId, version)
+      : removeVersionSpecificInformation(api);
+  });
+  return Promise.all(promises);
 }
